Add the ability to encrypt and decrypt strings
Provide a multimethod for encryption that accepts either strings
or byte-arrays.

Decryption is somewhat different in that a user might actually want
the byte-array versus the string. As such, this provides a new function
`decrypt-to-string` that will decrypt a token and convert it directly to
a string. The arguments to decrypt and decrypt-string are the same.

Tests have been added for new behaviour.